+++ src/main/java/drinkshop/repository/file/FileStocRepository.java
@@ -21,8 +21,8 @@ public class FileStocRepository
 
         int id = Integer.parseInt(elems[0]);
         String ingredient = elems[1];
-        int cantitate = Integer.parseInt(elems[2]);
-        int stocMinim = Integer.parseInt(elems[3]);
+        double cantitate = Double.parseDouble(elems[2]);
+        double stocMinim = Double.parseDouble(elems[3]);
 
         return new Stoc(id, ingredient, cantitate, stocMinim);
     }
